requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
tikapi==3.0.7
typer==0.7.0
urllib3==1.26.12
